refactor(core): Bring current executions into executions controller
