+++ cypress/e2e/20-workflow-executions.cy.ts
@@ -16,7 +16,7 @@ describe('Current Workflow Executions', () => {
 	it('should render executions tab correctly', () => {
 		createMockExecutions();
 		cy.intercept('GET', '/rest/executions?filter=*').as('getExecutions');
-		cy.intercept('GET', '/rest/executions-current?filter=*').as('getCurrentExecutions');
+		cy.intercept('GET', '/rest/executions/current?filter=*').as('getCurrentExecutions');
 
 		executionsTab.actions.switchToExecutionsTab();
 
@@ -34,7 +34,7 @@ describe('Current Workflow Executions', () => {
 
 	it('should not redirect back to execution tab when request is not done before leaving the page', () => {
 		cy.intercept('GET', '/rest/executions?filter=*');
-		cy.intercept('GET', '/rest/executions-current?filter=*');
+		cy.intercept('GET', '/rest/executions/current?filter=*');
 
 		executionsTab.actions.switchToExecutionsTab();
 		executionsTab.actions.switchToEditorTab();
@@ -63,7 +63,7 @@ describe('Current Workflow Executions', () => {
 		};
 
 		cy.intercept('GET', '/rest/executions?filter=*', throttleResponse);
-		cy.intercept('GET', '/rest/executions-current?filter=*', throttleResponse);
+		cy.intercept('GET', '/rest/executions/current?filter=*', throttleResponse);
 
 		executionsTab.actions.switchToExecutionsTab();
 		executionsTab.actions.switchToEditorTab();
